Should fix tests by moving the models to fixtures.
